[RAM-4] Move Logic/Data files into Logic/State
