+++ Assets/Source/Logic/Events/EventCommand.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using Source.Logic.Data;
+using Source.Logic.State;
 using Source.Utility;
 
 namespace Source.Logic.Events
