+++ Assets/Source/Logic/Events/MoveUnitsInDirectionEventCommand.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Source.Logic.Data;
+using Source.Logic.State;
 using Source.Utility;
 
 namespace Source.Logic.Events
