+++ Assets/Source/Logic/Events/UnitDeathEventCommand.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text;
-using Source.Logic.Data;
+using Source.Logic.State;
 using Source.Utility;
 using UnityEngine;
