+++ Assets/Source/Logic/State/BattlefieldStorage.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Source.Logic.Data
+namespace Source.Logic.State
 {
     [Serializable]
     public class BattlefieldStorage
